+++ activitysampling-domain/src/main/java/de/muspellheim/activitysampling/domain/Activities.java
@@ -1,0 +1,12 @@
+package de.muspellheim.activitysampling.domain;
+
+import java.time.*;
+import java.util.*;
+
+public interface Activities {
+  List<Activity> findAll();
+
+  List<Activity> findInPeriod(LocalDate from, LocalDate to);
+
+  void append(Activity activity);
+}