+++ activitysampling-domain/src/main/java/de/muspellheim/activitysampling/domain/Activity.java
@@ -3,9 +3,10 @@ package de.muspellheim.activitysampling.domain;
 import java.time.*;
 import java.util.*;
 
-public record Activity(LocalTime time, String description) {
+public record Activity(LocalDateTime timestamp, Duration duration, String description) {
   public Activity {
-    Objects.requireNonNull(time, "time");
+    Objects.requireNonNull(timestamp, "timestamp");
+    Objects.requireNonNull(duration, "duration");
     Objects.requireNonNull(description, "description");
   }
 }